// src/app/model/committes.ts
export class Committes {
      id:number =0;
      status:string ='';
      Scopeofthecommittee:string ='';
      Theadministrationconcerned: string ='';
      City: string ='';
      datastart: number=0 ;
      dataend: number =0;
      Typeofcommittee: any ='';
      Generalpurposeofthecommittee: string ='';
      numbercommittee:Number =0;
      name:string =''
}